==> 3_merge_datasets.ipynb <==
# About this notebook

This notebook does the following...
- Read climate and rice data
- Merge them and outputs `merged_data.csv`

import pandas as pd

climate_df = pd.read_csv('./climateipynb_output.csv', names=['Date', 'Rainfall', 'Max Temperature', 'MinTemperature', 'Relative Humidity', 'Wind Speed'], header=0)

rice_df = pd.read_csv('./riceipynb_output1.csv', names=['Date', 'Irrigated Palay Production', 'Rainfed Palay Production'], header=0)

climate_rice_df = pd.merge(climate_df, rice_df,  
                   on='Date',  
                   how='outer')

climate_rice_df.to_csv('merged_data.csv', index=False)

datasets = pd.read_csv('merged_data.csv')
datasets.head()
