==> 7_ricepred_irrigated.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from pycaret.regression import *
from pycaret.time_series import *

climate_forecast = pd.read_csv('climateforecast.csv', names=['Date', 'Rainfall', 'Max Temperature', 'Min Temperature', 'Relative Humidity', 'Wind Speed'], header=0)
climate_forecast.head()
dates = climate_forecast['Date']
climate_forecast.set_index('Date', inplace=True)

# Including Quarter
# climate_forecast['Quarter'] = pd.to_datetime(climate_forecast.index).quarter

irrice_final_model = load_model('irrice_final_model')

irrice_forecast = irrice_final_model.predict(climate_forecast)

irrice_forecast_df = pd.DataFrame(irrice_forecast)

irrice_forecast_df[0] = irrice_forecast_df[0].round()


irrice_forecast_df = irrice_forecast_df.assign(Date=dates)
irrice_forecast_df['Date'] = pd.to_datetime(irrice_forecast_df['Date'])

irrice_forecast_df.head()


plot_quarter = irrice_forecast_df.resample('Q', on='Date').mean()


plot_quarter.head()

plot_quarter.plot()

q = climate_forecast.assign(Irrigated = irrice_forecast)
q['Irrigated'] = q['Irrigated'].apply(lambda x: '{:.10f}'.format(x))
q.head()


plot_quarter.head()

q.head()

plot_quarter.plot(y=0, use_index=True, figsize=(15, 6), title='Quarterly Forecast of Irrigated Rice Production', grid=True, legend=False)
